pyproject.toml
[project]
name = "uploadpicker"
version = "0.1.0"
description = "Local GPU-ready hybrid image search tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.4.0",
  "sqlmodel>=0.0.22",
  "pillow>=10.4.0",
  "numpy>=1.26.4",
  "torch>=2.4.0",
  "open-clip-torch>=2.26.1",
  "faiss-cpu>=1.8.0.post1",
  "opencv-python>=4.10.0.84",
  "ultralytics>=8.3.0"
]

[project.optional-dependencies]
gpu = [
  "faiss-gpu>=1.7.2"
]
packaging = [
  "pyinstaller>=6.10.0"
]

[project.scripts]
uploadpicker = "app.runner:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
